% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{com}
\alias{com}
\title{Return complementary nucleotide}
\usage{
com(x)
}
\arguments{
\item{x}{nucleotide (A, T, G or C)}
}
\value{
complementary nucleotide
}
\description{
Return complementary nucleotide
}
