% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{draw.signature}
\alias{draw.signature}
\title{Draw component spectrum}
\usage{
draw.signature(sign, n.sign = NULL)
}
\arguments{
\item{sign}{vector of mutational component spectrum}

\item{n.sign}{name of component to show in legend}
}
\description{
Draw component spectrum
}
