% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{extract.comp}
\alias{extract.comp}
\title{Infer mutational components}
\usage{
extract.comp(rate, n.comp, fun = "kur")
}
\arguments{
\item{rate}{matrix of mutation rates of mutation types in genomic windows}

\item{n.comp}{number of components}
}
\value{
matrices of components spectra and intensities
}
\description{
Infer mutational components
}
