% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{parse.context}
\alias{parse.context}
\title{Parsing coded mutation type
 E.g. converts 'TCT_A' to a vector c('TCT','T','C>A','T')}
\usage{
parse.context(sign.names)
}
\arguments{
\item{sign.names}{vector of mutation types}
}
\value{
matrix of parsed mutation types
}
\description{
Parsing coded mutation type
 E.g. converts 'TCT_A' to a vector c('TCT','T','C>A','T')
}
