% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{plot.intensities}
\alias{plot.intensities}
\title{Draw component intensities along the genome}
\usage{
\method{plot}{intensities}(track, info, chr = 1, span.wind = 20,
  start = NULL, end = NULL)
}
\arguments{
\item{track}{vector of component intensities}

\item{info}{data.frame containing chr and start/end information about intensities}

\item{chr}{chromosome to show}

\item{span.wind}{number of windows to use for loess smoothing}

\item{start}{start of the region to show}

\item{end}{end of the region to show}
}
\description{
Draw component intensities along the genome
}
