% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{plot.reflection.matrix}
\alias{plot.reflection.matrix}
\title{Plot heatmap of refelction correlation between all pairs of components}
\usage{
\method{plot}{reflection.matrix}(icM, par = FALSE)
}
\arguments{
\item{icM}{matrix of spectra of mutational components}

\item{par}{boolean, TRUE if external par() is used. TRUE by default}
}
\value{
heatmap
}
\description{
Plot heatmap of refelction correlation between all pairs of components
}
