% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{reflection.scatter}
\alias{reflection.scatter}
\title{Scatterplot of spectrum and reverse complementary spectrum for two selected components}
\usage{
reflection.scatter(i, j, icM, par = FALSE)
}
\arguments{
\item{i}{index of first component}

\item{j}{index of second component, spectrum of which will be reverse complemented}

\item{icM}{matrix of spectra of mutational components}

\item{par}{boolean, TRUE if external par() is used. TRUE by default}
}
\value{
scatterplot
}
\description{
Scatterplot of spectrum and reverse complementary spectrum for two selected components
}
