% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{reflection.test}
\alias{reflection.test}
\title{Classify components and group in processes based on reflection property}
\usage{
reflection.test(icM, cutoff = 0.8)
}
\arguments{
\item{icM}{matrix of spectra of mutational components}

\item{cutoff}{cutoff on reflection property}
}
\value{
matrices of components classification and processes annotation in strand-independent/strand-dependent/noise/ambiguous
}
\description{
Classify components and group in processes based on reflection property
}
