% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{revert.context}
\alias{revert.context}
\title{Return reverse complementary mutation types}
\usage{
revert.context(sign.names)
}
\arguments{
\item{sign.names}{vector of mutation types}
}
\value{
vector of reverse complementary mutation types
}
\description{
Return reverse complementary mutation types
}
