% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{revert.matrix}
\alias{revert.matrix}
\title{Reverse parsed matrix of mutation type}
\usage{
revert.matrix(parse.matrix)
}
\arguments{
\item{parse.matrix}{matrix of parsed vector of mutation types}
}
\value{
matrix of reversed parsed mutation types
}
\description{
Reverse parsed matrix of mutation type
}
