% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{show.optimal.comp}
\alias{show.optimal.comp}
\title{Show number of real mutational components and processes depending on number of extracted components}
\usage{
show.optimal.comp(stat.extract, par = FALSE)
}
\arguments{
\item{stat.extract}{matrix output of select.ncomponents()}
}
\description{
Show number of real mutational components and processes depending on number of extracted components
}
