% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemut_functions.R
\name{visualize.bootstrap}
\alias{visualize.bootstrap}
\title{Estimate robustness of components spectra using bootstrap}
\usage{
visualize.bootstrap(boot, icM = NULL, par = FALSE)
}
\arguments{
\item{boot}{matrix of components spectra}

\item{icM}{matrix of components spectra. If supplied than reflection correlations are also shown.}

\item{par}{boolean. If TRUE, external plot par() configuration is used. FALSE by default}
}
\value{
boxplot of bootstrap correlations.
}
\description{
Estimate robustness of components spectra using bootstrap
}
